pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-agent-local-rag"
version = "0.1.0"
description = "Local multi-agent RAG system with LangGraph, Chroma, Neo4j, and web fallback"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "langchain>=0.3.0",
  "langgraph>=0.2.0",
  "langchain-community>=0.3.0",
  "langchain-text-splitters>=0.3.0",
  "langchain-chroma>=0.2.0",
  "chromadb>=0.5.5",
  "neo4j>=5.24.0",
  "langchain-openai>=0.2.0",
  "langchain-ollama>=0.2.0",
  "pypdf>=5.0.0",
  "Pillow>=10.0.0",
  "pytesseract>=0.3.10",
  "python-dotenv>=1.0.1",
  "ddgs>=8.0.0",
  "httpx>=0.27.0",
  "orjson>=3.10.0",
  "tenacity>=9.0.0",
  "rank-bm25>=0.2.2",
  "sentence-transformers>=3.0.1",
  "python-multipart>=0.0.9"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
